requirements.txt
Django==4.1.5
pillow
django-colorfield==0.8.0
django-crispy-forms==1.14.0

pylint==2.15.10